add some eindex

+++ views/LoginComponent.php
@@ -27,6 +27,8 @@ class LoginComponent
             $table_name,
             array(
                 'token' => $hash,
+                'browser' => $_SERVER['HTTP_USER_AGENT'],
+                'ip' => $_SERVER['REMOTE_ADDR'],
                 'created_at' => current_time('mysql', 1)
             )
         );
@@ -38,6 +40,8 @@ class LoginComponent
             <img src="<?php echo $qr; ?>" alt="<?php echo $url; ?>">
             </p>
             <br />
+            <p>Scan this QR with any logged in browser or browse link given bellow.</p>
+            <p><a href="javascript:;" class="<?php echo $this->instance->plugin_slug; ?>" data-url="<?php echo $url; ?>">Click Here to Copy Link</a></p>
             <br />
 
     <?php
@@ -47,6 +51,14 @@ class LoginComponent
     {
         global $wpdb;
         $table_name = $wpdb->prefix . rtrim($this->instance->plugin_prefix, "_");
+
+        # count number of rows
+
+        $count = $wpdb->get_var("SELECT COUNT(*) FROM $table_name WHERE created_at < '" . date('Y-m-d H:i:s', strtotime('-5 minutes')) . "'");
+
+        if ($count < 1) {
+            return;
+        }
         $deleteData = $wpdb->query(
             $wpdb->prepare(
                 "DELETE FROM $table_name WHERE created_at < %s",
